requirements.txt
pandas==2.2.2
numpy==2.1.1
scikit-learn==1.5.2
xgboost==2.1.1
statsmodels==0.14.3
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
textblob==0.18.0.post0
symspellpy==6.7.7
nltk==3.9.1
spacy==3.7.5
keybert==0.8.5
bertopic==0.16.2
sentence-transformers==3.0.1
transformers==4.44.2
torch==2.4.1
protobuf==5.28.2
streamlit==1.38.0
plotly==5.24.1
matplotlib==3.9.2
wordcloud==1.9.3
reportlab==4.2.2
fpdf2==2.7.9
schedule==1.2.2
beautifulsoup4==4.12.3
selenium==4.25.0
scrapy==2.11.2
googletrans==4.0.0rc1
langdetect==1.0.9
python-dotenv==1.0.1
fastparquet==2024.5.0
uvicorn==0.30.6
